Fix bad PerfView collection paramter
We use the /NoRundown switch on stop.
This is not meant for general use, and causes the modules of methods to not
be resolved as well as methods (which means that source code lookup also
fails.

I also limited the collection to a circular buffer of 1 GB by default.
This avoids creating very large files which are generally not useful.
ALso increased the bufferSize to 1GB to avoid lost events (this scenario
CAN produce alot of events).