Add OptimizationPreference=Speed NativeAOT runs
This is now a documented option: https://learn.microsoft.com/en-us/dotnet/core/deploying/native-aot/optimizing

I've put it on the same schedule as PGO runs since it's a similar thing but I don't know the motivation behind it. Should I tweak the modulus remainder to a different number than 1?